+++ client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.13.4]
+
+- Updated `avail-light-core` to 1.2.13
+
 ## [1.13.3](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.13.3) - 2025-10-21
 
 - Added log to tracing adapter